+++ changelog.md
@@ -377,3 +377,27 @@
     * get_farms_v2
     * get_farm_v3
     * get_farms_v3
+
+06-12-2021
+* Inventory Class
+    * get_application_pools_v3
+    * get_application_pool_v3
+    * get_machines_v2
+    * get_machine_v2
+    * get_global_desktop_entitlements_v2
+    * get_global_desktop_entitlement_v2
+    * get_global_application_entitlements_v2
+    * get_global_application_entitlement_v2
+
+13-12-2021
+* Inventory Class
+    * Added new_desktop_pool
+
+24-12-2021
+* Inventory Class
+    * Added desktop_validate_installed_applications
+    * Added farm_validate_installed_applications
+    * Added desktop_validate_specified_names
+    * Added farm_add_rds_servers
+    * Added farm_cancel_scheduled_maintenance
+    * Added farm_remove_rds_servers